ASADT3 | Release Version 1.0.6
Update ASADT version to 1.0.6 and add new features and improvements

This commit updates ASADT version to 1.0.6 with some major updates and improvements. It removes the support for the majority of tools to allow a new wave of more advanced tool execution. It also adds new features such as a tooltip view for each input value and command generation based on user input. Additionally, it updates the README file and adds a new image file. Finally, it includes instructions for users to update their version of ASADT and resources for reporting any issues or requesting new features.